Commit Strider Mountain source files.
Persistent bugs:

- Inability steer left or right when driving a car.
To bypass this, try the following:

- Enter in a car if not already.
- Save the game.
- Load your last save.

This is a temporary solution until properly fixed.

Fixed:

- "Could not load library client." when launching the mod.
-  Scene files not playing.
-  In-game menu.
-  Gameinfo.txt now has correct parameters

Changes:

- Map 1_sm_off_course

Added extra output to switch camera to continue through
introductory sequence.

This is used to overcome a particular bug where the APC
would not start forward, thus not reaching it's final
destination, which is responsible for triggering next
event relay.

- Map 7_sm_ascent

Divided terrains to prevent a crash caused by
too large displacements "Bad vertex data".

- Map 17_sm_bonus

Adjusted APC vehicle to face direction to the center of
the bridge.

This is used to overcome a particular bug where the APC
would not move to the middle of the bridge in the
last section of the level.