Commit City 7 source files.
Fixed:

- "Could not load library client." when launching the mod.
-  Scene files not playing.
-  In-game menu.
-  Gameinfo.txt now has correct parameters
-  Crash on maps dundas_square_f and yonge_st_f caused by destroying an
APC.
-  Crash on map st_michaels_f3_f due to invalid assault points.

Changes:

- Map st_michaels_f3_f

Removed first assault goal to prevent a crash.

- Map yonge_st_f

Added an extra relay entity to force the opening of combine gate located
at the end of the level. This new relay will trigger one minute after
entering the last area.

This is used to prevent a bug where citizens carrying RPGs would
not be able to fire rounds at the mortar, which is responsible for
opening the gate and allow players to destroy the APC, which is needed
to complete the level.