Commit Human Error source files.
Persistent bugs:

- Map he01_02
Random crash during the assault.

To overcome this, try the following:

- Save your game frequently. Reload the game in case of crash.

This is a temporary solution until properly fixed.

- Map he03_02
Noah does not walk inside the last room, prior to the final battle.

To bypass this, try the following:

- Simply push Noah toward the room, which will make him start the
required schedule
once inside.

This is a temporary solution until properly fixed.

Fixed:

- "Could not load library client." when launching the mod.
-  Gameinfo.txt now has correct parameters.
-  Missing materials.

Improvements:

- Improved impact effects for stunstick weapon.