Commit Triage source files.
Persistent bugs:

- Map chapter2b
Rare random crash in the last level section, prior to
canal entrance.

To overcome this, try the following:

- Save your game frequently. Reload the game in case of crash.

This is a temporary solution until properly fixed.

- Map chapter2e
Rare random crash in the middle of the second part of the level,
prior to the zone where players activate radio to tell rebels on
the other side to destroy it.

To bypass this, try the following:

- Save your game frequently. Reload the game in case of crash.

This is a temporary solution until properly fixed.

- Weapons not correctly drawn after map transition.

To bypass this, try the following:

- Toggle ironsights

This is a temporary solution until properly fixed.

Fixed:

- "Could not load library client." when launching the mod.
-  Scene files not playing.
-  Gameinfo.txt now has correct parameters.
-  Ryan door bust choreography not firing properly on map chapter1a.
-  Zombies not rendering correctly on map chapter2e.
-  End game video not starting after advisor encounter on map chapter2e.

Changes:

- Map chapter1a
- Added extra output to force Ryan sequence to start.

This is used to overcome bugs where logical entities would
fail to trigger sequences based on NPC kill count.

- Added a RPG with and a RPG round ammo crate on 4th floor.

This is used to overcome AI lead behaviors overriding RPG attack
schedules,
which caused Ryan to constantly move away and toward hiding spot on 4th
floor,
and not attempting to fire rockets at distracted boss chopper.

To bring back gameplay balance, RPG is stripped from players' equipment
upon
chopper destruction, which leaves them to pick up their previous weapon
before they traded for RPGs.

- Map chapter2e
- Adjusted zombie NPCs spawn positions.

This is used to prevent a bug where zombies would become invisible if
clipping
through displacement.

Improvements:

- Added English Closed caption (subtitles).