Commit "1187: Episode One" source files.
Persistent bugs:

- Weapons not correctly drawn after map transition.

To bypass this, try the following:

- Toggle ironsights

This is a temporary solution until properly fixed.

Fixed:

- "Could not load library client." when launching the mod.
-  Scene files not playing.
-  Gameinfo.txt now has correct parameters.
-  Fixed a bug that would cause the game to duplicate NPCs such
as Chris; John; Mike;

Changes:

- Removed Manual from game menu.

- Map 1187d1
Modified level flow to prevent players from accessing a location that
would
cause the game to crash (The nature of this bug is unknown).

To get to the basement, where several houndeyes are sleeping, players
should
use an airvent located in the small room on the right of entry, inside
apartment 126.

- Map 1187d2
Blocked hallway on second floor to prevent players from accessing a
location
that would cause the game to crash (The nature of this bug is unknown).

This change did not affect level flow, as players are not required to
pass
through this area.

Improvements:

- Added English Closed caption (subtitles).