Commit "School Adventures" source files.
Fixed:

- "Could not load library client." when launching the mod.
-  In-game menu.
-  Gameinfo.txt now has correct parameters.
-  Missing widescreen background texture.
-  Train not stopping, on map sa_04.
-  Crash upon entering final section (rebels waiting for teleporter), on
map sa_04.

Changes:

- Map sa_04

Added an additional output to stop the train to allow players to proceed
to the next stage.

This is used to overcome a situation where players would be unable to
exhaust manhack spawner, due to arbitrary manhack AI, which is
responsible
for summoning train arrival and stage completion.